==> NewMexico/WaterUse_PublicSupply/0_PreProcessNMSSPublicSupplyWaterUseData.ipynb <==
# Pre-processing New Mexico Site Specific Public Supply Water Use data for WaDE upload.
Date Updated: 10/11/2022

# working with data
import os
import numpy as np
import pandas as pd
import geopandas as gpd # the library that lets us read in shapefiles

# visulizaiton
import matplotlib.pyplot as plot
import seaborn as sns

# Cleanup
from datetime import datetime
pd.set_option('display.max_columns', 999)  # How to display all columns of a Pandas DataFrame in Jupyter Notebook

# Working Directory and Input File
workingDir = "G:/Shared drives/WaDE Data/NewMexico/SS_PublicSupplyWaterUse/RawInputData"

os.chdir(workingDir)

## Inputs and Dataframe Creation

# POD site with timeseries data
fileInput = "NMPWSPoints_input.csv"
dfpd = pd.read_csv(fileInput)

# WaDE UUID tracker for data assessment
if 'WaDEUUID' not in dfpd:
    dfpd['WaDEUUID'] = "nmPD" + dfpd.index.astype(str)
    dfpd.to_csv('NMPWSPoints_input.csv', index=False)

print(len(dfpd))
dfpd.head(1)

# Only looking at "Active" records.
dfpd = dfpd[dfpd['DWB_STATUS'] == 'A']
print(len(dfpd))
dfpd['DWB_STATUS'].unique()

# POU site data (only)
fileInput = "nm_pws_input.csv"
dfpu = pd.read_csv(fileInput)

# WaDE UUID tracker for data assessment
if 'WaDEUUID' not in dfpu:
    dfpu['WaDEUUID'] = "nmPU" + dfpu.index.astype(str)
    dfpu.to_csv('nm_pws_input.csv', index=False)

print(len(dfpu))
dfpu.head(1)

## POD Time Series Data
- 4 datasets between 2010, 2015 SW & GW.

# F2010_GW_AFY
dfpd_10GW = pd.DataFrame(index=dfpd.index)

# Variable Info
dfpd_10GW['in_VariableCV'] = "Withdrawal"
dfpd_10GW['in_VariableSpecificCV'] = ""

# Water Source Info
dfpd_10GW['in_WaterSourceTypeCV'] = "Groundwater" #change here

# Site Info
dfpd_10GW['in_County'] = dfpd['CountyName_HA']
dfpd_10GW['in_HUC8'] = dfpd['HUC_8']
dfpd_10GW['in_Latitude'] = dfpd['Latitude']
dfpd_10GW['in_Longitude'] = dfpd['Longitude']
dfpd_10GW['in_PODorPOUSite'] = "POD"
dfpd_10GW['in_SiteName'] = "Unspecified"
dfpd_10GW['in_SiteNativeID'] = dfpd['ID']

# Site VariableAmounts Info
dfpd_10GW['in_Amount'] = dfpd['F2010_GW_AFY'] #change here
dfpd_10GW['in_BeneficialUseCategory'] = "Unspecified"
dfpd_10GW['in_PopulationServed'] = dfpd['F2010_Population'] #change here
dfpd_10GW['in_ReportYearCV'] = "2010" #change here
dfpd_10GW['in_TimeframeEnd'] = "12/31/" + dfpd_10GW['in_ReportYearCV'].astype(str)
dfpd_10GW['in_TimeframeStart'] = "01/01/" + dfpd_10GW['in_ReportYearCV'].astype(str)

# link to site data
dfpd_10GW['linkKey'] = dfpd['Public_Water_System_Name_2019']

print(len(dfpd_10GW))
dfpd_10GW.head(1)

# F2010_SW_AFY
dfpd_10SW = pd.DataFrame(index=dfpd.index)

# Variable Info
dfpd_10SW['in_VariableCV'] = "Withdrawal"
dfpd_10SW['in_VariableSpecificCV'] = ""

# Water Source Info
dfpd_10SW['in_WaterSourceTypeCV'] = "Surface Water" #change here

# Site Info
dfpd_10SW['in_County'] = dfpd['CountyName_HA']
dfpd_10SW['in_HUC8'] = dfpd['HUC_8']
dfpd_10SW['in_Latitude'] = dfpd['Latitude']
dfpd_10SW['in_Longitude'] = dfpd['Longitude']
dfpd_10SW['in_PODorPOUSite'] = "POD"
dfpd_10SW['in_SiteName'] = "Unspecified"
dfpd_10SW['in_SiteNativeID'] = dfpd['ID']

# Site VariableAmounts Info
dfpd_10SW['in_Amount'] = dfpd['F2010_SW_AFY'] #change here
dfpd_10SW['in_BeneficialUseCategory'] = "Unspecified"
dfpd_10SW['in_PopulationServed'] = dfpd['F2010_Population'] #change here
dfpd_10SW['in_ReportYearCV'] = "2010" #change here
dfpd_10SW['in_TimeframeEnd'] = "12/31/" + dfpd_10GW['in_ReportYearCV'].astype(str)
dfpd_10SW['in_TimeframeStart'] = "01/01/" + dfpd_10GW['in_ReportYearCV'].astype(str)

# link to site data
dfpd_10SW['linkKey'] = dfpd['Public_Water_System_Name_2019']

print(len(dfpd_10SW))
dfpd_10SW.head(1)

# F2015_GW_AFY
dfpd_15GW = pd.DataFrame(index=dfpd.index)

# Variable Info
dfpd_15GW['in_VariableCV'] = "Withdrawal"
dfpd_15GW['in_VariableSpecificCV'] = ""

# Water Source Info
dfpd_15GW['in_WaterSourceTypeCV'] = "Groundwater" #change here

# Site Info
dfpd_15GW['in_County'] = dfpd['CountyName_HA']
dfpd_15GW['in_HUC8'] = dfpd['HUC_8']
dfpd_15GW['in_Latitude'] = dfpd['Latitude']
dfpd_15GW['in_Longitude'] = dfpd['Longitude']
dfpd_15GW['in_PODorPOUSite'] = "POD"
dfpd_15GW['in_SiteName'] = "Unspecified"
dfpd_15GW['in_SiteNativeID'] = dfpd['ID']

# Site VariableAmounts Info
dfpd_15GW['in_Amount'] = dfpd['F2015_GW_AFY'] #change here
dfpd_15GW['in_BeneficialUseCategory'] = "Unspecified"
dfpd_15GW['in_PopulationServed'] = dfpd['F2015_Population'] #change here
dfpd_15GW['in_ReportYearCV'] = "2015" #change here
dfpd_15GW['in_TimeframeEnd'] = "12/31/" + dfpd_15GW['in_ReportYearCV'].astype(str)
dfpd_15GW['in_TimeframeStart'] = "01/01/" + dfpd_15GW['in_ReportYearCV'].astype(str)

# link to site data
dfpd_15GW['linkKey'] = dfpd['Public_Water_System_Name_2019']

print(len(dfpd_15GW))
dfpd_15GW.head(1)

# F2015_SW_AFY
dfpd_15SW = pd.DataFrame(index=dfpd.index)

# Variable Info
dfpd_15SW['in_VariableCV'] = "Withdrawal"
dfpd_15SW['in_VariableSpecificCV'] = ""

# Water Source Info
dfpd_15SW['in_WaterSourceTypeCV'] = "Surface Water" #change here

# Site Info
dfpd_15SW['in_County'] = dfpd['CountyName_HA']
dfpd_15SW['in_HUC8'] = dfpd['HUC_8']
dfpd_15SW['in_Latitude'] = dfpd['Latitude']
dfpd_15SW['in_Longitude'] = dfpd['Longitude']
dfpd_15SW['in_PODorPOUSite'] = "POD"
dfpd_15SW['in_SiteName'] = "Unspecified"
dfpd_15SW['in_SiteNativeID'] = dfpd['ID']

# Site VariableAmounts Info
dfpd_15SW['in_Amount'] = dfpd['F2015_SW_AFY'] #change here
dfpd_15SW['in_BeneficialUseCategory'] = "Unspecified"
dfpd_15SW['in_PopulationServed'] = dfpd['F2015_Population'] #change here
dfpd_15SW['in_ReportYearCV'] = "2015" #change here
dfpd_15SW['in_TimeframeEnd'] = "12/31/" + dfpd_15SW['in_ReportYearCV'].astype(str)
dfpd_15SW['in_TimeframeStart'] = "01/01/" + dfpd_15SW['in_ReportYearCV'].astype(str)

# link to site data
dfpd_15SW['linkKey'] = dfpd['Public_Water_System_Name_2019']

print(len(dfpd_15SW))
dfpd_15SW.head(1)

# Concatenate POD Data Together
frames = [dfpd_10GW, dfpd_10SW, dfpd_15GW, dfpd_15SW]
dfpd_out = pd.concat(frames).reset_index(drop=True)
print(len(dfpd_out))

## POU Site Data
- just to show POU data for the POD timeseries above.

# left-join POD to POU data
dfpu = pd.merge(dfpu, dfpd, left_on='Wt_S_ID', right_on='WaterSystem_ID', how='left')
print(len(dfpu))
dfpu.head(1)

#Place of Use dataframe
dfpu_out = pd.DataFrame(index=dfpu.index)

# Variable Info
dfpu_out['in_VariableCV'] = "Withdrawal"
dfpu_out['in_VariableSpecificCV'] = ""

# Water Source Info
dfpu_out['in_WaterSourceTypeCV'] = "Unspecified"

# Site Info
dfpu_out['in_County'] = dfpu['CountyName_HA'] # from POD
dfpu_out['in_HUC8'] = dfpu['HUC_8_x']
dfpu_out['in_Latitude'] = dfpu['cent_Lat']
dfpu_out['in_Longitude'] = dfpu['cent_Long']
dfpu_out['in_PODorPOUSite'] = "POU"
dfpu_out['in_SiteName'] = dfpu['PblcSyN']
dfpu_out['in_SiteNativeID'] = dfpu['Wt_S_ID']

# Site VariableAmounts Info
dfpu_out['in_Amount'] = ""
dfpu_out['in_BeneficialUseCategory'] = "Unspecified"
dfpu_out['in_PopulationServed'] = ""
dfpu_out['in_ReportYearCV'] = ""
dfpu_out['in_TimeframeEnd'] = ""
dfpu_out['in_TimeframeStart'] = ""

# link to site data
dfpu_out['linkKey'] = dfpu['PblcSyN']

print(len(dfpu_out))
dfpu_out.head(1)

## Concatenate POD with POU data Site Data

frames = [dfpd_out, dfpu_out]
dfout = pd.concat(frames).reset_index(drop=True)
print(len(dfout))

## WaDE Custom Elements (due to missing info)

# Creating WaDE Custom VariableSpecificCV
# ----------------------------------------------------------------------------------------------------

def createVariableSpecificCV(inV, inBU, inWST):
    inV = str(inV).strip()
    inBU = str(inBU).strip().title()
    inWST = str(inWST).strip()
    
    outString = inV + "_Annual_" +  inBU + "_" + inWST
    
    return outString

dfout['in_VariableSpecificCV'] = dfout.apply(lambda row: createVariableSpecificCV(row['in_VariableCV'], 
                                                                                     row['in_BeneficialUseCategory'],
                                                                                     row['in_WaterSourceTypeCV']), axis=1)
dfout['in_VariableSpecificCV'].unique()

# Creating WaDE Custom water source native ID for easy water source identification
# ----------------------------------------------------------------------------------------------------

# Create temp WaterSourceNativeID dataframe of unique water source.
def assignWaterSourceNativeID(colrowValue):
    string1 = str(colrowValue)
    outstring = "WaDNMD_WS" + string1
    return outstring

dfWaterSourceNativeID = pd.DataFrame()
dfWaterSourceNativeID['in_WaterSourceTypeCV'] = dfout['in_WaterSourceTypeCV']
dfWaterSourceNativeID = dfWaterSourceNativeID.drop_duplicates()

dftemp = pd.DataFrame(index=dfWaterSourceNativeID.index)
dftemp["Count"] = range(1, len(dftemp.index) + 1)
dfWaterSourceNativeID['in_WaterSourceNativeID'] = dftemp.apply(lambda row: assignWaterSourceNativeID(row['Count']), axis=1)

# ----------------------------------------------------------------------------------------------------

# Retreive WaDE Custom water source native ID
def retrieveWaterSourceNativeID(A):
    if (A == '') or (pd.isnull(A)):
        outList = ''
    else:
        ml = dfWaterSourceNativeID.loc[(dfWaterSourceNativeID['in_WaterSourceTypeCV'] == A), 'in_WaterSourceNativeID']
        if not (ml.empty):  # check if the series is empty
            outList = ml.iloc[0]
        else:
            outList = ''
    return outList

dfout['in_WaterSourceNativeID'] = dfout.apply(lambda row: retrieveWaterSourceNativeID(row['in_WaterSourceTypeCV']), axis=1)
dfout['in_WaterSourceNativeID'].unique()

# if native site ID is not provdied / fill in blanks
# Creating WaDE Custom site native ID for easy site identification
# row.name = dataframe index of that row
# ----------------------------------------------------------------------------------------------------

def retrieveSiteNativeID(val, valIndex):
    if (val == '') or (pd.isnull(val)):
        outString = "WaDNMD_S" + str(valIndex)
    else:
        outString = val
    return outString

dfout['in_SiteNativeID'] = dfout.apply(lambda row: retrieveSiteNativeID(row['in_SiteNativeID'], row.name), axis=1)
dfout['in_SiteNativeID'].unique()

## Cleaning Output
- checking & changing data type & format.

# Convert History Year to YYYY-MM-DD format.

dfout['in_TimeframeEnd'] = pd.to_datetime(dfout['in_TimeframeEnd'], errors = 'coerce')
dfout['in_TimeframeEnd'] = pd.to_datetime(dfout["in_TimeframeEnd"].dt.strftime('%m/%d/%Y'))

dfout['in_TimeframeStart'] = pd.to_datetime(dfout['in_TimeframeStart'], errors = 'coerce')
dfout['in_TimeframeStart'] = pd.to_datetime(dfout["in_TimeframeStart"].dt.strftime('%m/%d/%Y'))

dfout.head(1)

# Converting numbers that are in string to float.

# in_Latitude & in_Longitude
dfout['in_Latitude'] = pd.to_numeric(dfout['in_Latitude'], errors='coerce')
dfout['in_Longitude'] = pd.to_numeric(dfout['in_Longitude'], errors='coerce')

# in_Amount
dfout['in_Amount'] = pd.to_numeric(dfout['in_Amount'], errors='coerce')

#in_ReportYearCV
# having some issues converting this to an int
dfout['in_ReportYearCV'] = pd.to_numeric(dfout['in_ReportYearCV'], errors='coerce')
dfout['in_ReportYearCV'] = dfout['in_ReportYearCV'].fillna(0).astype('int64')

#in_PopulationServed
# having some issues converting this to an int
dfout['in_PopulationServed'] = pd.to_numeric(dfout['in_PopulationServed'], errors='coerce')
dfout['in_PopulationServed'] = dfout['in_PopulationServed'].fillna(0).astype('int64')

dfout.head(1)

## Shapefile Data
- For attaching gemetry to csv inputs.

# PoU Shapefile Data
# Shapefile input
dfPoUshapetemp = gpd.read_file('shapefile/nm_pws_areas/nm_pws.shp')
print(len(dfPoUshapetemp))
dfPoUshapetemp.head(3)

columnsList = ['in_SiteNativeID', 'geometry']
dfPoUshape = pd.DataFrame(columns=columnsList)
dfPoUshape['in_SiteNativeID'] = dfPoUshapetemp['Wt_S_ID'].astype(str)
dfPoUshape['geometry'] = dfPoUshapetemp['geometry']
dfPoUshape = dfPoUshape.drop_duplicates(subset=None, keep='first', inplace=False, ignore_index=False)
dfPoUshape.head(3)

## Export Outputs

with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    print(dfout.dtypes)

#Exporting to Finished File
dfout.to_csv('P_nmSSPWMain.csv', index=False)  # The output
dfPoUshape.to_csv('P_nmSSPWGeometry.csv', index=False) # The output geometry.
