==> NewMexico/WaterUse_PublicSupply/7_NMssps_PODSiteToPOUSiteRelationships.ipynb <==
## Create podsitetopousiterelationships.csv file
Last Updated: 11/15/2022

Purpose: To create UT site specific public supply site POD and POU relation information and populate dataframe for WaDEQA 2.0.

Notes:    
- read in main processed input file, sites, and sitespecificamounts.
- separate out sites from POU to POD
- attach time info from sitespecificamounts to both POU and POD site dataframes, then attach linking element from main processed input fill.
- Attach POU dataframe to POD datafame if a common linking element.

#Needed Libararies

# working with data
import os
import numpy as np
import pandas as pd
import geopandas as gpd

# visulizaiton
import matplotlib.pyplot as plt
import seaborn as sns

# API retrieval
import requests
import json

# Cleanup
from datetime import datetime
pd.set_option('display.max_columns', 999)  # How to display all columns of a Pandas DataFrame in Jupyter Notebook

# Working Directory
workingDir = "G:/Shared drives/WaDE Data/NewMexico/SS_PublicSupplyWaterUse"
os.chdir(workingDir)

### Input Files

# Read in main processed input file
fileInput = "RawInputData/P_nmSSPWMain.csv"
dfmp = pd.read_csv(fileInput)
print(len(dfmp))
dfmp.head(1)

# Read in sites file
fileInput = "ProcessedInputData/Sites.csv"
dfs = pd.read_csv(fileInput)
print(len(dfs))
dfs.head(1)

# Read in sitespecificamounts file
fileInput = "ProcessedInputData/sitespecificamounts.csv"
dfsa = pd.read_csv(fileInput)
dfsa = dfsa.assign(SiteUUID=dfsa['SiteUUID'].str.split(',')).explode('SiteUUID').reset_index(drop=True)
print(len(dfsa))
dfsa.head(1)

### Extract Data

# extract unique in_SiteNativeID and linkKey from main processed input file
dfmp2 = dfmp[['in_SiteNativeID', 'linkKey']]
dfmp2 = dfmp2.dropna().drop_duplicates().reset_index(drop=True)
print(len(dfmp2))
dfmp2.head(2)

#create POU site file
dfspou = dfs[['SiteUUID', 'SiteNativeID', 'PODorPOUSite']]
dfspou = dfspou[dfspou['PODorPOUSite'] == 'POU']
dfspou = dfspou.dropna().drop_duplicates().reset_index(drop=True)
print(len(dfspou))
dfspou.head(2)

#create POD site file
dfspod = dfs[['SiteUUID', 'SiteNativeID', 'PODorPOUSite']]
dfspod = dfspod[dfspod['PODorPOUSite'] == 'POD']
dfspou = dfspou.dropna().drop_duplicates().reset_index(drop=True)
print(len(dfspou))
dfspou.head(2)

# extract useable site info with StartDate & EndDate from sitespecificamounts file
dfsa2 = dfsa[['SiteUUID', 'TimeframeStart', 'TimeframeEnd']]
dfsa2 = dfsa2.dropna().drop_duplicates().reset_index(drop=True)
print(len(dfsa2))
dfsa2.head(2)

# left join POU sites to sitespecificamounts to main processed input file
dfmergeU = pd.DataFrame(index=dfspou.index)

dfmergeU = pd.merge(dfspou, dfsa2, on='SiteUUID', how='left')
dfmergeU = pd.merge(dfmergeU, dfmp2, left_on='SiteNativeID', right_on='in_SiteNativeID', how='left')
dfmergeU = dfmergeU.drop_duplicates().reset_index(drop=True)
print(len(dfmergeU))
dfmergeU.head(2)

# left join POD sites to sitespecificamounts to main processed input file
dfmergeD = pd.DataFrame(index=dfspod.index)

dfmergeD = pd.merge(dfspod, dfsa2, on='SiteUUID', how='left')
dfmergeD = pd.merge(dfmergeD, dfmp2, left_on='SiteNativeID', right_on='in_SiteNativeID', how='left')
dfmergeD = dfmergeD.drop_duplicates().reset_index(drop=True)
print(len(dfmergeD))
dfmergeD.head(2)

# Left Join POD file with POU file with matching linking elements. Drop NaN rows
dfmergeDU = pd.DataFrame(index=dfmergeD.index)

dfmergeDU = pd.merge(dfmergeD, dfmergeU, on='linkKey', how='left')
dfmergeDU = dfmergeDU.dropna(subset=['SiteUUID_y']).drop_duplicates().reset_index(drop=True)
print(len(dfmergeDU))
dfmergeDU.head(2)

### Create Output Dataframe

columnslist = ["PODSiteUUID", "POUSiteUUID", "StartDate", "EndDate"]
outdf = pd.DataFrame(index=dfmergeDU.index, columns=columnslist)  # The output dataframe
print(len(outdf))
outdf.head(1)

outdf['PODSiteUUID'] = dfmergeDU['SiteUUID_x']
outdf['POUSiteUUID'] = dfmergeDU['SiteUUID_y']
outdf['StartDate'] = dfmergeDU['TimeframeStart_x']
outdf['EndDate'] = dfmergeDU['TimeframeEnd_x']
outdf = outdf.dropna().drop_duplicates().reset_index(drop=True).sort_values(by=['PODSiteUUID', 'POUSiteUUID', 'StartDate'])
print(len(outdf))
outdf.head()

# The working output DataFrame for WaDE 2.0 input.
if not outdf.empty:
    outdf.to_csv('ProcessedInputData/podsitetopousiterelationships.csv', index=False)
